+++ NEWS.md
@@ -1,3 +1,9 @@
+# spbabel dev
+
+
+* spbabel<- replacement function now drops attributes if object and row numbers
+not the same
+
 # spbabel 0.4.0
 
 * new function 'map_table' to produce the more general multiple-table model